+++ module/Application/src/Controller/AdminSeatController.php
@@ -4,6 +4,7 @@ namespace Application\Controller;
 
 use Application\Entity\Event;
 use Application\Entity\Seat;
+use Application\Service\FirebaseService;
 use Doctrine\ORM\EntityManager;
 use Zend\Http\Request;
 use Zend\Mvc\Controller\AbstractActionController;
@@ -55,6 +56,15 @@ class AdminSeatController extends AbstractActionController
                     ])) {
                 $entityManager->remove($seat);
                 $entityManager->flush();
+
+                /** @var FirebaseService $firebaseService */
+                $firebaseService = $this->getServiceManager()->get(FirebaseService::class);
+                $firebaseService->update($event->getId(),[
+                    $data['seatNumber'] => [
+                        'seatNumber' => $data['seatNumber'],
+                        'status' => Seat::AVAILABLE
+                    ]
+                ]);
 
                 $response = [
                     'statusCode' => 200,
