+++ module/Application/view/application/index/view.phtml
@@ -1,12 +1,14 @@
 <div class="alert alert-success success-message response-message hidden">
-    <button type="button" class="close" data-dismiss="alert" aria-label="Close"><span aria-hidden="true">&times;</span></button>
+    <button type="button" class="close" data-dismiss="alert" aria-label="Close"><span aria-hidden="true">&times;</span>
+    </button>
     <h3>Sua reserva foi gerada com sucesso. Agora você deverá confirmar a mesma.</h3>
     <small>No mundo real, um email seria enviado com o link de confirmação, mas para este teste
         será apenas um redirecionamento em <strong>10s</strong>
     </small>
 </div>
 <div class="alert alert-danger error-message response-message hidden">
-    <button type="button" class="close" data-dismiss="alert" aria-label="Close"><span aria-hidden="true">&times;</span></button>
+    <button type="button" class="close" data-dismiss="alert" aria-label="Close"><span aria-hidden="true">&times;</span>
+    </button>
     <h3>Não foi possível completar sua reserva.</h3>
     <span class="error-message-content"></span>
 </div>
@@ -35,7 +37,8 @@
             if (array_key_exists($i, $selectedSeats)) {
                 $complementClass = $selectedSeats[$i]['class'] . ' unclickable';
             } ?>
-            <div class="seat <?php echo $complementClass; ?>" id="seat-<?php echo $i; ?>">&nbsp;<?php echo str_pad($i, 3, '0', STR_PAD_LEFT); ?>&nbsp;
+            <div class="seat <?php echo $complementClass; ?>" id="seat-<?php echo $i; ?>">
+                &nbsp;<?php echo str_pad($i, 3, '0', STR_PAD_LEFT); ?>&nbsp;
             </div>
         <?php endfor; ?>
         <br clear="all"/>
@@ -43,28 +46,64 @@
         <h5>Legenda</h5>
         <span class="your-choice">Sua escolha</span>&nbsp;
         <span class="pre-reserved">Pré reserva</span>&nbsp;
-        <span class="just-selected">Recém selecionada</span>&nbsp;
-        <span class="unavailable">Reserva confirmada</span>
+        <span class="unavailable">Reserva confirmada</span>&nbsp;
+        <span class="seat-color">Poltrona livre</span>&nbsp;
         <br clear="all"/>
-        <br />
-    </div>
-    <div class="col-xs-6 col-md-4 event-details">
-        <h3><?php echo $this->event->getName(); ?></h3>
         <br/>
+    </div>
+    <div class="col-xs-12 col-md-4 event-details">
+        <h3><?php echo $this->event->getName(); ?></h3>
+        R$<?php echo number_format($this->event->getTicketAmount(), 2, ',', '.'); ?>
+        <br/><br/>
         Local: <?php echo $this->event->getLocation(); ?>
         <br/>
         Capacidade: <?php echo $this->event->getCapacity(); ?> expectadores
         <br/>
         Data: <?php echo $this->event->getShowDate()->format('d/m/Y H:i'); ?>
         <hr>
-        <h4>Sua escolha</h4>
+        <h4>Seu pedido</h4>
         <div class="seats-selected"></div>
         <hr>
         <label>Seu e-mail</label>
         <input type="email" placeholder="Email" id="email" class="form-control">
         <br/>
         <input type="hidden" id="eventId" value="<?php echo $this->event->getId(); ?>">
+        <input type="hidden" id="ticketAmount" value="<?php echo $this->event->getTicketAmount(); ?>">
+        Total do pedido: <strong>R$<span id="orderAmount">0,00</span></strong>
         <a href="#" class="btn btn-success" id="btn-action">Reservar</a>
     </div>
 </div>
 <br clear="all"/>
+
+<?php
+$this->inlineScript()->captureStart();
+echo <<<JS
+jQuery(document).ready(function ($) {
+    var ref = firebase.database().ref('demo/events/{$this->event->getId()}/seats');
+    
+    ref.on('child_changed', function (snapshot) {
+        var data = snapshot.val();
+        var seat = $('#seat-' + data.seatNumber);
+        
+        $(seat).removeClass('pre-reserved');
+        $(seat).removeClass('unavailable');
+        $(seat).removeClass('unclickable');
+
+        var itemClass = '';
+        switch (data.status) {
+            case 0:
+                itemClass = 'pre-reserved unclickable';
+                break;
+            case 1:
+                itemClass = 'unavailable unclickable';
+                break;
+            default:
+                itemClass = '';
+                break;
+        }
+
+        $(seat).addClass(itemClass);
+    });
+});
+JS;
+$this->inlineScript()->captureEnd();
