+++ module/Application/view/layout/admin.phtml
@@ -24,6 +24,7 @@
     <?= $this->headScript()
         ->prependFile($this->basePath('js/jquery.priceformat.min.js'))
         ->prependFile($this->basePath('js/jquery.datetimepicker.full.js'))
+        ->prependFile($this->basePath('js/admin.js'))
         ->prependFile($this->basePath('js/app.js'))
         ->prependFile($this->basePath('js/bootstrap.min.js'))
         ->prependFile($this->basePath('js/jquery-3.1.0.min.js'))
